requirements.txt
numpy==1.12.1
joblib==0.13.2
matplotlib==2.2.3
object_detection.egg==info
pandas==0.23.4
astropy==1.3.3
Pillow==6.0.0
config==0.4.2
object_detection==0.0.3
tensorflow==1.14.0
